% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{fq_spec}
\alias{fq_spec}
\title{Analyzer specification}
\usage{
fq_spec()
}
\description{
Abstract parent of the analyzer specs. A spec records which analyzer to run
and its hyperparameters. It holds no fitted state. Construct a concrete spec
with a constructor such as \code{fq_kmeans()}.
}
